# 03-rest-apis-that-dont-suck/docs/openapi.yaml
openapi: 3.0.3
info:
  title: LaunchPad API
  description: API for the LaunchPad startup discovery platform.
  version: 1.0.0
  contact:
    name: LaunchPad Team

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      description: Check if the server is running.
      tags:
        - System
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time

  /api/startups:
    get:
      summary: List startups
      description: >
        Returns a paginated list of startups. Supports sorting,
        filtering by category, and text search.
      tags:
        - Startups
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Items per page
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, name, upvotes]
            default: created_at
          description: Sort field
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
        - name: category
          in: query
          schema:
            type: integer
          description: Filter by category ID
        - name: search
          in: query
          schema:
            type: string
          description: Search in name and tagline
      responses:
        '200':
          description: List of startups with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Startup'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    post:
      summary: Create a startup
      description: Create a new startup listing.
      tags:
        - Startups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStartup'
      responses:
        '201':
          description: Startup created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Startup'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /api/startups/{id}:
    get:
      summary: Get a startup
      description: Get a single startup by ID.
      tags:
        - Startups
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Startup ID
      responses:
        '200':
          description: Startup found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Startup'
        '404':
          description: Startup not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

    put:
      summary: Update a startup
      description: Update an existing startup. Only include fields to change.
      tags:
        - Startups
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Startup ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStartup'
      responses:
        '200':
          description: Startup updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Startup'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Startup not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

    delete:
      summary: Delete a startup
      description: Delete a startup by ID.
      tags:
        - Startups
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Startup ID
      responses:
        '204':
          description: Startup deleted (no content)
        '404':
          description: Startup not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

  /api/categories:
    get:
      summary: List categories
      description: Returns all categories.
      tags:
        - Categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'

  /api/categories/{id}:
    get:
      summary: Get a category
      description: Get a single category by ID.
      tags:
        - Categories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Category ID
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Category'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

components:
  schemas:
    Startup:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: TechFlow AI
        tagline:
          type: string
          example: AI-powered workflow automation
        description:
          type: string
          example: TechFlow AI is a platform that...
        url:
          type: string
          format: uri
          example: https://techflow.ai
        category_id:
          type: integer
          example: 1
        upvotes:
          type: integer
          example: 42
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        category_name:
          type: string
          example: AI/ML

    CreateStartup:
      type: object
      required:
        - name
        - tagline
        - description
        - url
        - category_id
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: CloudDeploy
        tagline:
          type: string
          minLength: 10
          maxLength: 200
          example: Deploy anywhere in seconds with zero configuration
        description:
          type: string
          minLength: 50
          maxLength: 2000
          example: CloudDeploy is a deployment platform that automatically detects your framework and deploys it.
        url:
          type: string
          format: uri
          example: https://clouddeploy.dev
        category_id:
          type: integer
          minimum: 1
          example: 2

    UpdateStartup:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        tagline:
          type: string
          minLength: 10
          maxLength: 200
        description:
          type: string
          minLength: 50
          maxLength: 2000
        url:
          type: string
          format: uri
        category_id:
          type: integer
          minimum: 1

    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: AI/ML
        description:
          type: string
          example: Artificial Intelligence and Machine Learning
        created_at:
          type: string
          format: date-time

    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 47
        totalPages:
          type: integer
          example: 5

    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: name
              message:
                type: string
                example: Name is required

    NotFoundError:
      type: object
      properties:
        error:
          type: string
          example: Startup with id '999' not found